+++ src/main/java/blackjack/BlackjackApplication.java
@@ -1,19 +1,14 @@
 package blackjack;
 
-import blackjack.domain.gamer.Dealer;
-import blackjack.domain.gamer.Players;
+import blackjack.game.BlackjackGame;
 import blackjack.view.InputView;
 import blackjack.view.OutputView;
 
 public class BlackjackApplication {
 	public static void main(String[] args) {
-		BlackjackController blackjackController = new BlackjackController(InputView.getInstance(),
+		BlackjackGame blackjackGame = new BlackjackGame(InputView.getInstance(),
 			OutputView.getInstance());
 
-		Players players = blackjackController.createPlayers();
-		Dealer dealer = blackjackController.createDealer();
-
-		blackjackController.startGame(dealer, players);
-		blackjackController.printResult(dealer, players);
+		blackjackGame.start();
 	}
 }
